Refactored and reorganised test suite
