+++ ouroboros-network/ouroboros-network.cabal
@@ -247,14 +247,17 @@ library sim-tests-lib
   exposed-modules:
     Ouroboros.Network.BlockFetch.Examples
     Ouroboros.Network.MockNode
-    Test.Ouroboros.Network.BlockFetch
     Test.Ouroboros.Network.Diffusion.Node
     Test.Ouroboros.Network.Diffusion.Node.ChainDB
     Test.Ouroboros.Network.Diffusion.Node.MiniProtocols
     Test.Ouroboros.Network.Diffusion.Node.NodeKernel
     Test.Ouroboros.Network.Diffusion.Policies
-    Test.Ouroboros.Network.KeepAlive
+    Test.Ouroboros.Network.Diffusion.Testnet.Cardano
+    Test.Ouroboros.Network.Diffusion.Testnet.Cardano.Node
     Test.Ouroboros.Network.LedgerPeers
+    Test.Ouroboros.Network.MiniProtocols.BlockFetch
+    Test.Ouroboros.Network.MiniProtocols.KeepAlive
+    Test.Ouroboros.Network.MiniProtocols.TxSubmission
     Test.Ouroboros.Network.MockNode
     Test.Ouroboros.Network.Mux
     Test.Ouroboros.Network.NodeToClient.Version
@@ -271,9 +274,6 @@ library sim-tests-lib
     Test.Ouroboros.Network.PeerSelection.PublicRootPeers
     Test.Ouroboros.Network.PeerSelection.RootPeersDNS
     Test.Ouroboros.Network.PeerState
-    Test.Ouroboros.Network.Testnet
-    Test.Ouroboros.Network.Testnet.Simulation.Node
-    Test.Ouroboros.Network.TxSubmission
     Test.Ouroboros.Network.Version
 
 -- Simulation tests, and IO tests which don't require native system calls.
