+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/MiniProtocols/BlockFetch.hs
@@ -6,7 +6,7 @@
 
 {-# OPTIONS_GHC -Wno-incomplete-uni-patterns #-}
 
-module Test.Ouroboros.Network.BlockFetch
+module Test.Ouroboros.Network.MiniProtocols.BlockFetch
   ( PeerGSVT (..)
   , tests
   ) where
