+++ ouroboros-network/sim-tests-lib/Test/Ouroboros/Network/PeerSelection/PublicRootPeers.hs
@@ -19,6 +19,8 @@ import Ouroboros.Network.Testing.Utils (ShrinkCarefully, prop_shrink_nonequal,
 import Test.Ouroboros.Network.PeerSelection.Instances
 
 
+import Cardano.Node.PublicRootPeers (CardanoPublicRootPeers (..))
+import Cardano.Node.PublicRootPeers qualified as CNPRP
 import Ouroboros.Network.PeerSelection.PeerAdvertise (PeerAdvertise)
 import Ouroboros.Network.PeerSelection.PublicRootPeers (PublicRootPeers (..))
 import Test.QuickCheck
