+++ ouroboros-network/sim-tests/Main.hs
@@ -4,10 +4,12 @@ import Main.Utf8 (withUtf8)
 import Test.Tasty
 
 import Test.ChainProducerState qualified (tests)
-import Test.Ouroboros.Network.BlockFetch qualified (tests)
 import Test.Ouroboros.Network.Diffusion.Policies qualified (tests)
-import Test.Ouroboros.Network.KeepAlive qualified (tests)
+import Test.Ouroboros.Network.Diffusion.Testnet.Cardano qualified (tests)
 import Test.Ouroboros.Network.LedgerPeers qualified (tests)
+import Test.Ouroboros.Network.MiniProtocols.BlockFetch qualified (tests)
+import Test.Ouroboros.Network.MiniProtocols.KeepAlive qualified (tests)
+import Test.Ouroboros.Network.MiniProtocols.TxSubmission qualified (tests)
 import Test.Ouroboros.Network.MockNode qualified (tests)
 import Test.Ouroboros.Network.NodeToClient.Version qualified (tests)
 import Test.Ouroboros.Network.NodeToNode.Version qualified (tests)
@@ -20,8 +22,6 @@ import Test.Ouroboros.Network.PeerSelection.PeerMetric qualified
 import Test.Ouroboros.Network.PeerSelection.PublicRootPeers qualified
 import Test.Ouroboros.Network.PeerSelection.RootPeersDNS qualified
 import Test.Ouroboros.Network.PeerState qualified (tests)
-import Test.Ouroboros.Network.Testnet qualified (tests)
-import Test.Ouroboros.Network.TxSubmission qualified (tests)
 import Test.Ouroboros.Network.Version qualified (tests)
 
 main :: IO ()
@@ -34,24 +34,24 @@ tests =
   [ Test.ChainProducerState.tests
 
     -- network logic
-  , Test.Ouroboros.Network.Version.tests
-  , Test.Ouroboros.Network.PeerState.tests
-  , Test.Ouroboros.Network.BlockFetch.tests
-  , Test.Ouroboros.Network.PeerSelection.tests
+  , Test.Ouroboros.Network.Diffusion.Policies.tests
+  , Test.Ouroboros.Network.Diffusion.Testnet.Cardano.tests
+  , Test.Ouroboros.Network.LedgerPeers.tests
+  , Test.Ouroboros.Network.MiniProtocols.BlockFetch.tests
+  , Test.Ouroboros.Network.MiniProtocols.KeepAlive.tests
+  , Test.Ouroboros.Network.MiniProtocols.TxSubmission.tests
+  , Test.Ouroboros.Network.NodeToClient.Version.tests
+  , Test.Ouroboros.Network.NodeToNode.Version.tests
   , Test.Ouroboros.Network.PeerSelection.Json.tests
   , Test.Ouroboros.Network.PeerSelection.KnownPeers.tests
   , Test.Ouroboros.Network.PeerSelection.LocalRootPeers.tests
-  , Test.Ouroboros.Network.PeerSelection.PublicRootPeers.tests
   , Test.Ouroboros.Network.PeerSelection.MockEnvironment.tests
   , Test.Ouroboros.Network.PeerSelection.PeerMetric.tests
+  , Test.Ouroboros.Network.PeerSelection.PublicRootPeers.tests
   , Test.Ouroboros.Network.PeerSelection.RootPeersDNS.tests
-  , Test.Ouroboros.Network.KeepAlive.tests
-  , Test.Ouroboros.Network.TxSubmission.tests
-  , Test.Ouroboros.Network.NodeToNode.Version.tests
-  , Test.Ouroboros.Network.NodeToClient.Version.tests
-  , Test.Ouroboros.Network.Testnet.tests
-  , Test.Ouroboros.Network.LedgerPeers.tests
-  , Test.Ouroboros.Network.Diffusion.Policies.tests
+  , Test.Ouroboros.Network.PeerSelection.tests
+  , Test.Ouroboros.Network.PeerState.tests
+  , Test.Ouroboros.Network.Version.tests
 
     -- pseudo system-level
   , Test.Ouroboros.Network.MockNode.tests
